+++ src/components/beta/gux-tag/tests/gux-tag.e2e.ts
@@ -101,6 +101,7 @@ describe('gux-tag-beta', () => {
   describe('a11y', () => {
     [
       'default',
+      'default-subtle',
       'navy',
       'blue',
       'electric-purple',