+++ src/components/stable/gux-advanced-dropdown/tests/gux-advanced-dropdown.e2e.ts
@@ -14,7 +14,7 @@ describe('gux-advanced-dropdown', () => {
     const page = await newSparkE2EPage({ html });
     const element = await page.find('gux-advanced-dropdown');
     await a11yCheck(page);
-    expect(element).toHaveClass('hydrated');
+    expect(element).toHaveAttribute('hydrated');
   });
 
   it('should open the dropdown on click', async () => {