+++ src/components/stable/gux-content-search/tests/gux-content-search.e2e.ts
@@ -159,7 +159,7 @@ describe('gux-content-search', () => {
       const inputSpy = await page.spyOnEvent('input');
 
       expect(clearButton).toBeNull();
-      expect(element).toHaveClass('hydrated');
+      expect(element).toHaveAttribute('hydrated');
 
       await input.press('KeyT');
       await input.press('KeyE');
@@ -211,7 +211,7 @@ describe('gux-content-search', () => {
       let currentMatch = await element.getProperty('currentMatch');
       let matchCount = await element.getProperty('matchCount');
 
-      expect(element).toHaveClass('hydrated');
+      expect(element).toHaveAttribute('hydrated');
       expect(clearButton).not.toBeNull();
 
       expect(value).toBe('TEST');
@@ -253,7 +253,7 @@ describe('gux-content-search', () => {
       );
       const inputSpy = await page.spyOnEvent('input');
 
-      expect(element).toHaveClass('hydrated');
+      expect(element).toHaveAttribute('hydrated');
       expect(clearButton).not.toBeNull();
       expect(nextButton).not.toBeNull();
       expect(previousButton).not.toBeNull();
@@ -454,7 +454,7 @@ describe('gux-content-search', () => {
       );
       const inputSpy = await page.spyOnEvent('input');
 
-      expect(element).toHaveClass('hydrated');
+      expect(element).toHaveAttribute('hydrated');
       expect(clearButton).not.toBeNull();
       expect(nextButton).not.toBeNull();
       expect(previousButton).not.toBeNull();
@@ -531,7 +531,7 @@ describe('gux-content-search', () => {
       );
       const inputSpy = await page.spyOnEvent('input');
 
-      expect(element).toHaveClass('hydrated');
+      expect(element).toHaveAttribute('hydrated');
       expect(clearButton).not.toBeNull();
       expect(nextButton).not.toBeNull();
       expect(previousButton).not.toBeNull();